// src/components/engineers/types.ts
export type Engineer = {
  id: string;
  name: string;
  discipline: string;
  organization: string;
  province: string;
  photo: string;
  shortBio: string;
  fullBio: string;
  career: string[];
  achievements: string[];
  projects: string[];
  awards: string[];
  badges: string[];
  linkedin: string;
};